+++ src/_test-helpers/dataFactories.js
@@ -1,6 +1,10 @@
 const { createObjectListFactoryFromSchema } = require('./utils');
 const constants = require('./constants');
-const generateEffectsList = createObjectListFactoryFromSchema({
+
+/**
+ * @type {function(numEntries: number, valueGetter?: (propName: string, index: number, defaultValue: any) => any,  objectExtender?: (obj: object, index: number) => object): import('../datamine-types').IProcEffect[]}
+ */
+const generateProcEffectsList = createObjectListFactoryFromSchema({
 	'effect delay time(ms)/frame': (index) => `delay-${index}`,
 	'proc id': (index) => `id-${index}`,
 	'unknown proc id': () => undefined,
@@ -9,6 +13,9 @@ const generateEffectsList = createObjectListFactoryFromSchema({
 	'target type': () => constants.ARBITRARY_TARGET_TYPE,
 });
 
+/**
+ * @type {function(numEntries: number, valueGetter?: (propName: string, index: number, defaultValue: any) => any,  objectExtender?: (obj: object, index: number) => object): import('../datamine-types').IBurstDamageFramesEntry[]}
+ */
 const generateDamageFramesList = createObjectListFactoryFromSchema({
 	'effect delay time(ms)/frame': (index) => `delay-${index}`,
 	'frame times': (index) => Array.from({ length: index + 1 }, (_, timeIndex) => timeIndex),
@@ -18,7 +25,19 @@ const generateDamageFramesList = createObjectListFactoryFromSchema({
 	hits: (index) => index + 1,
 });
 
+/**
+ * @type {function(numEntries: number, valueGetter?: (propName: string, index: number, defaultValue: any) => any,  objectExtender?: (obj: object, index: number) => object): { [id: string]: import('../buffs/buff-metadata').IPassiveMetadataEntry }}
+ */
+const generatePassiveMetadataObject = (...args) => createObjectListFactoryFromSchema({
+	ID: (index) => `${index}`,
+	Name: (index) => `passive-entry-name-${index}`,
+})(...args).reduce((acc, entry) => {
+	acc[entry.ID] = entry;
+	return entry;
+}, {});
+
 module.exports = {
 	generateDamageFramesList,
-	generateEffectsList,
+	generateProcEffectsList,
+	generatePassiveMetadataObject,
 };