+++ src/buffs/getNameForPassive.spec.js
@@ -1,43 +1,73 @@
 const testConstants = require('../_test-helpers/constants');
 const getNameForPassive = require('./getNameForPassive').default;
+const { PASSIVE_METADATA } = require('./buff-metadata');
 
 describe('getNameForPassive method', () => {
-	[
-		{
-			input: testConstants.ARBITRARY_PASSIVE_ID,
-			name: 'a valid passive id',
-			shouldHaveEntry: true,
-		},
-		{
-			input: testConstants.ARBITRARY_INVALID_PASSIVE_ID,
-			name: 'an invalid passive id',
-			shouldHaveEntry: false,
-		},
-		{
-			input: void 0,
-			name: 'an undefined id',
-			shouldHaveEntry: false,
-		},
-		{
-			input: null,
-			name: 'a null id',
-			shouldHaveEntry: false,
-		},
-	].forEach(testCase => {
-		it(`returns ${testCase.shouldHaveEntry ? 'a string' : 'an empty string'} for ${testCase.name}`, () => {
-			const result = getNameForPassive(testCase.input);
-			expect(typeof result === 'string')
-				.withContext('result is not a string')
-				.toBe(true);
-			if (testCase.shouldHaveEntry) {
-				expect(result.length)
-					.withContext('result is an empty etring')
-					.toBeGreaterThan(0);
-			} else {
-				expect(result.length)
-					.withContext('result is not an empty string')
-					.toBe(0);
-			}
+	const expectNonEmptyString = (result, expected) => {
+		// use 2 expects to guard against result or expected being an empty string
+		expect(result.length)
+			.withContext('result is an empty etring')
+			.toBeGreaterThan(0);
+		expect(result).toBe(expected);
+	};
+
+	describe('for invalid metadata inputs', () => {
+		[
+			{
+				name: 'is null',
+				value: null,
+			},
+			{
+				name: 'is not an object',
+				value: 'some value',
+			},
+		].forEach(testCase => {
+			it(`returns empty string if the metadata parameter ${testCase.name}`, () => {
+				const result = getNameForPassive(testConstants.KNOWN_ARBITRARY_PASSIVE_ID, testCase.value);
+				expect(result).toBe('');
+			});
 		});
+	});
+
+	describe('when a valid metadata object is passed in', () => {
+		const arbitraryId = 'arbitrary id';
+		const arbitraryMissingId = 'arbitrary missing id';
+
+		[
+			{
+				input: void 0,
+				name: 'an undefined id',
+			},
+			{
+				input: null,
+				name: 'a null id',
+			},
+			{
+				input: arbitraryMissingId,
+				name: 'a passive id not in the metadata',
+			},
+			{
+				input: arbitraryId,
+				name: 'a passive id without a Name',
+			},
+		].forEach(testCase => {
+			it(`returns an empty string for ${testCase.name}`, () => {
+				const metadata = { [arbitraryId]: { ID: arbitraryId } };
+				const result = getNameForPassive(testCase.input, metadata);
+				expect(result).toBe('');
+			});
+		});
+
+		it('returns the Name field is of a given metadata entry', () => {
+			const arbitraryName = 'arbitrary name';
+			const metadata = { [arbitraryId]: { ID: arbitraryId, Name: arbitraryName } };
+			const result = getNameForPassive(arbitraryId, metadata);
+			expectNonEmptyString(result, arbitraryName);
+		});
+	});
+
+	it('defaults to PASSIVE_METADATA when metadata is not specified', () => {
+		const result = getNameForPassive(testConstants.KNOWN_ARBITRARY_PASSIVE_ID);
+		expectNonEmptyString(result, PASSIVE_METADATA[testConstants.KNOWN_ARBITRARY_PASSIVE_ID].Name);
 	});
 });