+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] - YYYY-MM-DD
+### Δ Changes
+* buffs: The following methods now have an optional metadata parameters for their respective metadata sources.
+	* `getMetadataForPassive`
+	* `getMetadataForProc`
+	* `getNameForPassive`
+	* `getNameForProc`
+	* `isAttackingProcId`
+	* `getEffectName` (allows passing in both passive and proc metadata)
+* bursts: Add optional metadata parameter to `getExtraAttackDamageFramesEntry`.
 
 ## [0.7.0] - 2020-06-26
 ### ⚠ Breaking Changes